WIP Introwork to constraint and parsing for type checking
To do type checking, I am first pulling all of the statements
within the functoin that is being analyzed, and then generating
the constraints for each statement.

Once the constraints have all been generated, solving the constraints
with a unification algoirithm will be the next step.

Currently, constraining statements which are:
1. declarations without initializations
2. constant assignments
3. typevar assignments

have all been roughly completed -- unsure of how much more detail is needed
for these constraints (this will be found in the unification alg). The main
holding point for for the constrain portion of this problem right now
is dealing with the constraint of expressions (x = x + y + z + 2).

Additionally, this approach doesn't currently account for type casting.
Another assumption made (which works for ints, so long as there's no casting),
is that for operations, all of the types need to be equal -- but this isn't the case
for say double, which can handle "Double x  = 5.0 + z" where z is an int.