Bug fix removing multiple instances of the same type-variable term
Previously:
When creating the constraint system, each term was generated without regard to
any previous constraints.

This means, that [[x]] = int and [[x]] = [[y]] would create 4 different term
instances. This double-creation of the "x" term causes issues in unification.
When going to unify, even if you unify x with int, if you have a later constraint
that would require unification with double -- the script would permit this new unification
because it is working on a separate instance of x, one that hasn't been unified with int.

Now:
When creating constraints, specifically term creation within those constraints...
There is a check to see if the type-variable (IDTerm, or really just an identifier/variable)
has already been created as a term, if it has, then essentially the new constraint will
reference this previously-existing term.

This allows for correct unifications.

Small notes:
There are also changes to allow for an easier equality check on any term instance.
There are changes regarding constraint generation which preserves the type of the term,
making it easier for further processing in "Unify.py". The distinguishers are:
1. IDTerm -> Identifier (variable name)
2. TypeTerm -> Proper type